+++ lib/govuk_publishing_components/presenters/devolved_nations_helper.rb
@@ -15,15 +15,16 @@ module GovukPublishingComponents
         nation_keys.each_with_index.map { |k, i| name_for_position(k, i, nation_keys.count, use_english_translation) }.join
       end
 
-      def name_for_position(nation, position, total, use_english_translation)
+      def name_for_position(nation, position, total, use_english_translation, cy_initial_mutation: true)
         key = if position.zero?
-                :start
+                cy_initial_mutation ? :start : :middle
               elsif position == (total - 1)
                 :end
               else
                 :middle
               end
-        I18n.t("components.devolved_nations.#{nation}.#{key}", locale: use_english_translation ? :en : nil)
+        name = I18n.t("components.devolved_nations.#{nation}.#{key}", locale: use_english_translation ? :en : nil)
+        position.zero? || position == (total - 1) ? name : ", #{name}"
       end
 
       def ga4_applicable_nations_title_text(remove_connector_word = nil)
@@ -41,7 +42,7 @@ module GovukPublishingComponents
 
         url_to_nation_keys.transform_values do |nation_keys|
           nations_text = nation_keys.each_with_index.map { |key, index|
-            name_for_position(key, index, nation_keys.count, true)
+            name_for_position(key, index, nation_keys.count, false, cy_initial_mutation: false)
           }.join
 
           alternative_content_text(nations_text)